Move shell variable substitution to pkg
