+++ pkg/shellvar/template_test.go
@@ -1,4 +1,4 @@
-package template
+package shellvar
 
 import (
 	"testing"
